+++ packages/ub-pub/i18n.js
@@ -3,17 +3,9 @@ const _ = require('lodash')
 const __i18n = {
   monkeyRequestsDetected: 'Your request has been processed, but we found that it is repeated several times. Maybe you key fuse?'
 }
+const FORMAT_RE = /{(\d+)}/g
 
-/**
- * see docs in ub-pub main module
- * @private
- * @param {String} localeString
- * @returns {*}
- */
-module.exports.i18n = function i18n (localeString) {
-  let res = __i18n[localeString]
-  if (res !== undefined || localeString == null) return res
-
+function domainBasedLocalization (localeString) {
   // $App is accessible only inside adminUI
   if (typeof $App === 'undefined') return localeString
 
@@ -58,8 +50,29 @@ module.exports.i18n = function i18n (localeString) {
 /**
  * see docs in ub-pub main module
  * @private
+ * @param {String} localeString
+ * @param {...*} formatArgs Format args
+ * @returns {*}
+ */
+module.exports.i18n = function i18n (localeString, ...formatArgs) {
+  if (localeString == null) return localeString
+  let res = __i18n[localeString]
+  res = res || domainBasedLocalization(localeString)
+  if (formatArgs && formatArgs.length && (typeof res === 'string')) {
+    return res.replace(FORMAT_RE, function (m, i) {
+      return formatArgs[i]
+    })
+  } else {
+    return res
+  }
+}
+
+/**
+ * see docs in ub-pub main module
+ * @private
  * @param {Object} localizationObject
+ * @returns {Object} new i18n object
  */
 module.exports.i18nExtend = function i18nExtend (localizationObject) {
-  _.merge(__i18n, localizationObject)
+  return _.merge(__i18n, localizationObject)
 }